+++ src/Core/engine.cpp
@@ -94,8 +94,8 @@ void Engine::init(Window* window){
 	ZeroMemory(&m_viewport, sizeof(D3D11_VIEWPORT));
 	m_viewport.TopLeftX = 0;
 	m_viewport.TopLeftY = 0;
-	m_viewport.Width = win_props->width; // SCREEN_WIDTH
-	m_viewport.Height = win_props->height; // SCREEN_HEIGHT
+	m_viewport.Width = (float) win_props->width; // SCREEN_WIDTH
+	m_viewport.Height = (float) win_props->height; // SCREEN_HEIGHT
 	m_viewport.MinDepth = 0.0f;
 	m_viewport.MaxDepth = 1.0f;
